test(2025-12-06/2): optimize BuildIntArray functions with templates and memory pre-allocation
- Template-ize BuildIntArray functions to support uint8/16/32/64 types
- Move template implementations to test::wwjson and test::yyjson namespaces
- Add optional size parameter for wwjson RawBuilder pre-allocation
- Optimize yyjson API usage with proper uint/sint function calls
- Add --size command line parameter (int type) for memory estimation
- Move test::BuildJson to test::wwjson namespace with size support
- Update all test cases to use new template functions
- Add compare_array_output test to verify wwjson vs yyjson output consistency
- Clean up old function declarations from test_data.h/.cpp

+++ include/wwjson.hpp
@@ -176,7 +176,7 @@ struct GenericBuilder
     /* ---------------------------------------------------------------------- */
     
     /// Constructor with optional initial capacity.
-    GenericBuilder(int capacity = 1024) { json.reserve(capacity); }
+    GenericBuilder(size_t capacity = 1024) { json.reserve(capacity); }
 
     /// Copy constructor.
     GenericBuilder(const GenericBuilder& other) = default;
